+++ packages/util-crypto/src/base58/decode.spec.ts
@@ -7,6 +7,12 @@ import { base32Decode } from '../base32/index.js';
 import { base58Decode } from './index.js';
 
 describe('base58Encode', (): void => {
+  it('decodes an empty string)', (): void => {
+    expect(
+      base58Decode('')
+    ).toEqual(new Uint8Array());
+  });
+
   it('encodes a base58 to a base32', (): void => {
     expect(
       base58Decode('b2rhk6GMPQF3hfzwXTaNYFLKomMeC6UXdUt6jZKPpeVirLtV')